+++ src/components/Ring.tsx
