+++ src/index.tsx
@@ -1,7 +1,7 @@
 import { createRoot } from "react-dom/client";
 import "./scss/style.scss";
 
-import App from "./com/App";
+import App from "./App";
 
 const root = createRoot(document.getElementById("app"));
 root.render(<App />);
